Removed ViewListItemElement.
